core-api: split up routing hooks module
